Include RoleArn and invalid bad routes
- Improve ecs-server docs
- Returned credentials via ECS server now include the RoleARN
- Bad paths processed by the default route now return a 404

Fixes: #561
